Revert "Push switch statement logic down into TableType"
This reverts commit 6e0cb09d8e5577428b49e9cbddfdcaa5ffe71746.
