--- atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/TableType.java
@@ -1,80 +1,0 @@
-/**
- * Copyright 2015 Palantir Technologies
- *
- * Licensed under the BSD-3 License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.palantir.atlasdb.keyvalue.dbkvs.impl;
-
-import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.dbkvs.OracleDdlConfig;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.oracle.OracleOverflowQueryFactory;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.oracle.OracleOverflowWriteTable;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.oracle.OracleRawQueryFactory;
-
-public enum TableType {
-    RAW(3) {
-        @Override
-        protected DbQueryFactory getDbQueryFactory(TableReference tableRef, OracleDdlConfig config) {
-            return new OracleRawQueryFactory(DbKvs.internalTableName(tableRef), config);
-        }
-
-        @Override
-        public DbWriteTable getWriteTable(TableReference tableRef, ConnectionSupplier conns,
-                OracleDdlConfig config) {
-            return OracleOverflowWriteTable.create(DbKvs.internalTableName(tableRef), conns, config);
-        }
-    }, OVERFLOW(4) {
-        @Override
-        protected DbQueryFactory getDbQueryFactory(TableReference tableRef, OracleDdlConfig config) {
-            return new OracleOverflowQueryFactory(DbKvs.internalTableName(tableRef), config);
-        }
-
-        @Override
-        public DbWriteTable getWriteTable(TableReference tableRef, ConnectionSupplier conns,
-                OracleDdlConfig config) {
-            return new SimpleDbWriteTable(DbKvs.internalTableName(tableRef), conns, config);
-        }
-    };
-
-    private final int id;
-
-    TableType(int id) {
-        this.id = id;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public static TableType byId(int id) {
-        switch (id) {
-            case 3:
-                return RAW;
-            case 4:
-                return OVERFLOW;
-            default:
-                throw new IllegalArgumentException("Unknown table size: " + id);
-        }
-    }
-
-    public DbReadTable getReadTable(TableReference tableRef, ConnectionSupplier conns, OracleDdlConfig config) {
-        return new UnbatchedDbReadTable(conns, getDbQueryFactory(tableRef, config));
-    }
-
-    protected abstract DbQueryFactory getDbQueryFactory(TableReference tableRef, OracleDdlConfig config);
-
-    public abstract DbWriteTable getWriteTable(
-            TableReference tableRef,
-            ConnectionSupplier connectionSupplier,
-            OracleDdlConfig config);
-}
