+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/postgres/PostgresDdlTable.java
@@ -25,7 +25,7 @@ import com.palantir.atlasdb.keyvalue.dbkvs.PostgresDdlConfig;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionSupplier;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.DbDdlTable;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.DbKvs;
-import com.palantir.atlasdb.keyvalue.dbkvs.impl.TableType;
+import com.palantir.atlasdb.keyvalue.dbkvs.impl.TableSize;
 import com.palantir.exception.PalantirSqlException;
 import com.palantir.nexus.db.sql.AgnosticResultSet;
 import com.palantir.nexus.db.sql.ExceptionCheck;
@@ -72,7 +72,7 @@ public class PostgresDdlTable implements DbDdlTable {
                             "INSERT INTO %s (table_name, table_size) VALUES (?, ?)",
                             config.metadataTableName()),
                     tableName.getQualifiedName(),
-                    TableType.RAW.getId());
+                    TableSize.RAW.getId());
         }, ExceptionCheck::isUniqueConstraintViolation);
     }
